Rename responses -> commands
