+++ src/commands/cat.js
