+++ src/commands/emote.js
