+++ src/commands/fish.js
