+++ src/commands/info.js
