+++ src/commands/reboot.js
