+++ src/commands/substitute.js
