+++ src/commands/weather.js
