+++ src/commands/whois.js
