+++ src/commands/xkcd.js
